diplom part 2

+++ product/src/main/java/practicum/ru/product/category/CategoryServiceImpl.java
@@ -3,6 +3,8 @@ package practicum.ru.product.category;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import practicum.ru.product.dto.CategoryDto;
+import practicum.ru.product.event.EventRepository;
+import practicum.ru.product.exception.ConflictException;
 import practicum.ru.product.exception.NotFoundException;
 
 import java.util.List;
@@ -12,21 +14,29 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     @Autowired
-    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryMapper categoryMapper) {
+    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryMapper categoryMapper, EventRepository eventRepository) {
         this.categoryRepository = categoryRepository;
         this.categoryMapper = categoryMapper;
+        this.eventRepository = eventRepository;
     }
 
     @Override
     public Category createCategory(CategoryDto categoryDto) {
+        if (categoryRepository.findByName(categoryDto.getName()) != null) {
+            throw new ConflictException("категория с именем [" + categoryDto.getName() + "] уже существует");
+        }
         return categoryRepository.save(categoryMapper.toCategory(categoryDto));
     }
 
     @Override
     public Category updateCategory(Long catId, CategoryDto categoryDto) {
         Category category = getCategoryById(catId);
+        if (categoryRepository.findByName(categoryDto.getName()) != null) {
+            throw new ConflictException("категория с именем [" + categoryDto.getName() + "] уже существует");
+        }
         category.setName(categoryDto.getName());
         return categoryRepository.save(category);
     }
@@ -48,6 +58,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void deleteCategory(Long id) {
         getCategoryById(id);
+        if (!eventRepository.findByCategoryId(id).isEmpty()) {
+            throw new ConflictException("Категория привязана к событиям");
+        }
         categoryRepository.deleteById(id);
     }
 }
